+++ src/pupil_labs/invisible_lsl_relay/outlets.py
@@ -23,6 +23,7 @@ class PupilInvisibleOutlet:
         timestamp_query,
         outlet_name_prefix,
         outlet_uuid,
+        acquisition_info,
     ):
         self._outlet_uuid = outlet_uuid
         self._channels = channel_func()
@@ -32,6 +33,7 @@ class PupilInvisibleOutlet:
             outlet_type,
             outlet_format,
             outlet_name_prefix,
+            acquisition_info,
         )
         self._timestamp_query = timestamp_query
 
@@ -47,7 +49,7 @@ class PupilInvisibleOutlet:
 
 
 class PupilInvisibleGazeOutlet(PupilInvisibleOutlet):
-    def __init__(self, device_id, outlet_prefix=None):
+    def __init__(self, device_id, outlet_prefix=None, world_camera_serial=None):
         PupilInvisibleOutlet.__init__(
             self,
             channel_func=pi_gaze_channels,
@@ -56,11 +58,14 @@ class PupilInvisibleGazeOutlet(PupilInvisibleOutlet):
             timestamp_query=pi_extract_from_sample('timestamp_unix_seconds'),
             outlet_name_prefix=outlet_prefix,
             outlet_uuid=f'{device_id}_Gaze',
+            acquisition_info=set_acquisition_info(
+                version=VERSION, world_camera_serial=world_camera_serial
+            ),
         )
 
 
 class PupilInvisibleEventOutlet(PupilInvisibleOutlet):
-    def __init__(self, device_id, outlet_prefix=None):
+    def __init__(self, device_id, outlet_prefix=None, world_camera_serial=None):
         PupilInvisibleOutlet.__init__(
             self,
             channel_func=pi_event_channels,
@@ -69,20 +74,38 @@ class PupilInvisibleEventOutlet(PupilInvisibleOutlet):
             timestamp_query=pi_extract_from_sample('timestamp_unix_seconds'),
             outlet_name_prefix=outlet_prefix,
             outlet_uuid=f'{device_id}_Event',
+            acquisition_info=set_acquisition_info(
+                version=VERSION, world_camera_serial=world_camera_serial
+            ),
         )
 
 
 def pi_create_outlet(
-    outlet_uuid, channels, outlet_type, outlet_format, outlet_name_prefix
+    outlet_uuid,
+    channels,
+    outlet_type,
+    outlet_format,
+    outlet_name_prefix,
+    acquisition_info,
 ):
     stream_info = pi_streaminfo(
-        outlet_uuid, channels, outlet_type, outlet_format, outlet_name_prefix
+        outlet_uuid,
+        channels,
+        outlet_type,
+        outlet_format,
+        outlet_name_prefix,
+        acquisition_info,
     )
     return lsl.StreamOutlet(stream_info)
 
 
 def pi_streaminfo(
-    outlet_uuid, channels, type_name: str, channel_format, outlet_name_prefix
+    outlet_uuid,
+    channels,
+    type_name: str,
+    channel_format,
+    outlet_name_prefix,
+    acquisition_info,
 ):
     stream_info = lsl.StreamInfo(
         name=f"{outlet_name_prefix}_{type_name}",
@@ -91,7 +114,9 @@ def pi_streaminfo(
         channel_format=channel_format,
         source_id=outlet_uuid,
     )
-    stream_info.desc().append_child_value("pupil_invisible_lsl_relay_version", VERSION)
+    xml_acquisition = stream_info.desc().append_child("acquisition")
+    for key in acquisition_info.keys():
+        xml_acquisition.append_child_value(key, acquisition_info[key])
     xml_channels = stream_info.desc().append_child("channels")
     [chan.append_to(xml_channels) for chan in channels]
     return stream_info
@@ -99,3 +124,14 @@ def pi_streaminfo(
 
 def get_lsl_time_offset():
     return time.time() - lsl.local_clock()
+
+
+def set_acquisition_info(
+    version, world_camera_serial, manufacturer='Pupil Labs', model='Pupil Invisible'
+):
+    return {
+        'manufacturer': manufacturer,
+        'model': model,
+        'world_camera_serial': world_camera_serial,
+        'pupil_invisible_lsl_relay_version': version,
+    }