+++ src/pupil_labs/invisible_lsl_relay/relay.py
@@ -10,15 +10,26 @@ logger = logging.getLogger(__name__)
 
 
 class Relay:
-    def __init__(self, device_ip, device_port, device_identifier, outlet_prefix):
+    def __init__(
+        self,
+        device_ip,
+        device_port,
+        device_identifier,
+        outlet_prefix,
+        world_camera_serial,
+    ):
         self.device_ip = device_ip
         self.device_port = device_port
         self.receiver = DataReceiver(device_ip, device_port)
         self.gaze_outlet = outlets.PupilInvisibleGazeOutlet(
-            device_id=device_identifier, outlet_prefix=outlet_prefix
+            device_id=device_identifier,
+            outlet_prefix=outlet_prefix,
+            world_camera_serial=world_camera_serial,
         )
         self.event_outlet = outlets.PupilInvisibleEventOutlet(
-            device_id=device_identifier, outlet_prefix=outlet_prefix
+            device_id=device_identifier,
+            outlet_prefix=outlet_prefix,
+            world_camera_serial=world_camera_serial,
         )
         self.gaze_sample_queue = asyncio.Queue()
         self.publishing_gaze_task = None